Move decorator helpers to testing_utils.py
